Make the checker pattern blink
This brings the checker pattern closer to the behavior of the test
signal. Indeed a static pattern provides a light energy level of exactly
half a full steady-state white frame plus half a full steady-state black
frame (assuming no shenanigans like dynamic contrast, local dimming or
ABL). But the test signal is the opposite of steady-state, and its
average light energy level if highly influenced by the the transient
response of the system, especially at high frame rates. This commit
tries to approximate this case by making the checker pattern behave like
a whole forest of test signals, spread across the screen so that they
hopefully cancel each other out in the sensor reading.

The main downside of doing this is that if the various patterns do not
integrate well in the instrument (e.g. the instrument is too close to
the screen and only "sees" one square), the analyzer could mistake the
checker pattern for the test signal.